legacy/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-multipart==0.0.6
aiohttp==3.9.1
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
google-api-python-client==2.108.0 